pyproject.toml
[project]
name = "ai-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rich>=14.3.3",
]
